Some optimizations
Track the start address and current head and tail of the unconsumed
buffer array separately. This lets us simply repoint the head when our
sentinel is satisfied and we return data to the caller. The free/malloc
that we used to do was very expensive.

This might cause increased memory use, but to help guard against that we
repoint the head and tail to the start if the buffer is ever fully
consumed. We'll probably have to do a periodic memmove though for cases
where the buffer stays ahead of consumers, perhaps when (head - start)
crosses a certain threshhold.

NB I had this code "ready" the day before but there was a subtle memory
corruption bug I had to track down.