Try to prevent too much waste at buffer head
If there is room, the remaining data in the buffer will get shifted
forward to use already allocated space that contains data that was
already consumed.

Also added a `get_stats` function that returns a couple numbers. The
first is the number of times the operation described above has taken
place and the second number is the max size that has been allocated for
a buffer.