A couple API tweaks to aid extensibility
The Connection class now has a RECV_FLAGS class attribute that can be
used to pass custom flags to recv(2).

The Service class has a new do_accept() method that accepts a new
connection and returns the socket and address in a tuple. This allows
custom Service subclasses to define what accepting means for them.

It also has a ConnectionClass attribute (defaults to Connection) for
subclasses that want to use a different type of connection.